requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0
django-crispy-forms==2.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.3
ecdsa==0.18.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jwcrypto==1.5.0
oauthlib==3.2.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wrapt==1.16.0
